update test case and add new method

+++ run-test.sh
@@ -16,12 +16,19 @@
 trap 'echo "Ctrl+C pressed. Exiting..."; exit 1' INT
 
 # Programs to run 
-programs=("python3 main.py") # TODO: Add the programs to run here
+programs=("python3 main.py" "python3 main2.py" "python3 main3.py") # TODO: Add the programs to run here
 
 # Input Folder Name
 input_folder="test/other-input"
 output_folder="test-out"
 valid_output_folder="test/other-output"
+
+# check if the --validate flag is set
+if [ "$1" = "--validateonly" ]; then
+    # run the python script to validate the output
+    python3 validate.py $input_folder $output_folder
+    exit 0
+fi
 
 # Delete all the files in the output folder
 rm -rf $output_folder
@@ -32,12 +39,13 @@ test_files=$(ls $input_folder)
 # Timeout duration in seconds (maximum time a program is allowed to run per test case)
 timeout_duration=300
 
+
 # Initialize the arrays to store the statistics
 program_pass_count=()
+program_improve_count=()
 program_total_count=()
 program_timeout_count=()
 program_time=()
-
 
 # Get the OS type for OS specific commands
 os_type=$(uname)
@@ -51,6 +59,7 @@ do
     # get the start time and initialize the iteration variables
     start_time=$(date +"%s")
     pass_count=0
+    improve_count=0
     total_count=0
 
     # clean the conetents of the output folder
@@ -97,6 +106,7 @@ do
         elif [ $output -lt $valid_output ]; then
             echo "     [PASS-BETTER]"
             pass_count=$((pass_count+1))
+            improve_count=$((improve_count+1))
         else
             echo "     [FAIL]"
         fi
@@ -117,27 +127,73 @@ do
     # get the end time and save the results
     end_time=$(date +"%s")
     program_pass_count+=($pass_count)
+    program_improve_count+=($improve_count)
     program_total_count+=($total_count)
     program_time+=($((end_time - start_time)))
     echo "Passed $pass_count out of $total_count tests in $((end_time - start_time)) seconds" 
 
 done
 
-echo "----------------------------------------------------------------------------------"
-echo "|                               Test Suite Summary                               |"
-echo "----------------------------------------------------------------------------------"
-printf "| %-20s | %6s | %6s | %8s | %14s | %7s |\n" "Program" "Passed" "Total" "Percentage" "Duration (sec)" "Timeout"
-echo "----------------------------------------------------------------------------------"
+echo "---------------------------------------------------------------------------------------------"
+echo "|                                    Test Suite Summary                                     |"
+echo "---------------------------------------------------------------------------------------------"
+printf "| %-20s | %6s | %8s | %6s | %8s | %14s | %7s |\n" "Program" "Passed" "Improved" "Total" "Percentage" "Duration (sec)" "Timeout"
+echo "---------------------------------------------------------------------------------------------"
 
 for program in "${programs[@]}"
 do
     index=$(printf "%s\n" "${programs[@]}" | grep -n -m 1 "$program" | cut -d: -f1)
     pass=${program_pass_count[$index-1]}
+    improve=${program_improve_count[$index-1]}
     total=${program_total_count[$index-1]}
     duration=${program_time[$index-1]}
     timeout=${program_timeout_count[$index-1]}
-    # Program | Passed | Total | Percentage | Duration | Timeout
-    printf "| %-20s | %6d | %6d | %9.2f%% | %14.2f | %7d |\n" "${program:0:20}" $pass $total $(echo "scale=2; $pass / $total * 100" | bc) $duration $timeout
+    # Program | Passed | Improved | Total | Percentage | Duration | Timeout
+    printf "| %-20s | %6d | %8d | %6d | %9.2f%% | %14.2f | %7d |\n" "${program:0:20}" $pass $improve $total $(echo "scale=2; $pass / $total * 100" | bc) $duration $timeout
 done
 
-echo "----------------------------------------------------------------------------------"
+echo "---------------------------------------------------------------------------------------------"
+
+# Run the python script to validate the output
+if [ "$1" = "--validate" ]; then
+    echo "Running the validation script"
+    python3 validate.py $input_folder $output_folder
+
+    # check if the update flag is set
+    if [ "$2" = "--update" ]; then
+        echo "Updating the valid output folder"
+        update_count=0
+        total_precent_improvement=0
+        # for each file in the output folder, check if it is an improvement and update the valid output folder
+        for file in $test_files
+        do
+            precent_improvement=0
+            output_file=$(echo $file | sed 's/.in/.out.valid/')
+            # if the output file contains the .valid extension, continue
+            if [[ $output_file == *".valid"* ]]; then
+                # Get the number of removals from the valid output (first line of the file)
+                valid_output=$(head -n 1 $valid_output_folder/$output_file)
+                valid_output=$(echo $valid_output | sed 's/[^0-9]*//g')
+
+                # Get the number of removals from the output file (first line of the file)
+                output=$(head -n 1 $output_folder/$output_file)
+                output=$(echo $output | sed 's/[^0-9]*//g')
+
+                # Check if the output is valid (if equal pass, if less then pass-better, if more then fail)
+                if [ $output -lt $valid_output ]; then
+                    echo "     [UPDATE] $output_file is an improvement"
+                    cp $output_folder/$output_file $valid_output_folder/$output_file
+                    update_count=$((update_count+1))
+                    precent_improvement=$(echo "scale=2; ($valid_output - $output) / $valid_output * 100" | bc)
+                    total_precent_improvement=$(echo "scale=2; $total_precent_improvement + $precent_improvement" | bc)
+                fi
+            fi
+        done
+        echo "Updated $update_count files"
+        if [ $update_count -gt 0 ]; then
+            echo "Average improvement: $(echo "scale=2; $total_precent_improvement / $update_count" | bc)%"
+        fi
+        
+    fi
+    echo "Validation complete"
+fi
